Prepare the code base for `php-cs-fixer`
We are using Matt Allan's Laravel Code Style, which sets the coding
style to conform to laravel's standards. I have included this package in
the "require" and not "require-dev" because this package will use this
to properly format the files it generates.

Notice also the `hooks/pre-commmit.php` file, which should be copied (or
linked) into `.git/hooks/pre-commit`. This file formats php files about
to be committed.