Stash unstaged changes before `php-cs-fix`ing
On commit, we have a hook to fix the code style of PHP files. If one
stages a file, and changes it further without staging it, then on commit
the full changes would be added to the staging area. To avoid this, we
stash all non-staged changes, apply the fix, and only then pop the top
stash.