Major overhaul of the code
The `TableInformation` was a bit of a mess, and exposed Doctrine details
that I think are better left abstracted away. This overhaul changed
essentially the `ColumnFaker` and `RuleGenerator` classes (and their
tests!), although some other generators also had to be changed to
conform to the new surface of the `TableInformation` class.