Change the vaidation of boolean fields
We were using the `boolean`  validation rule, which validates that the
value can be cast as boolean (1, 'true', etc.), but HTML sends a `on`
string associated with checkboxes and does not even send the field for
unselected checkboxes. As such, all boolean values in the database must
be nullable at the request level.