Validate boolean fields against the 'boolean' rule
HTML works like this: a form ith a checkbox sends the value if that
checkbox is selected (if no value attribute exists, the default value
'on' is used), or does not send the key/value pair at all if the
checkbox is not selected. This means that boolean values have a special
and annoying treatment. Also, validation is weird, because there is no
difference between sending a false or not sending the value at all.

To solve this, we will use the "standard" workaround of having a hidden
field on the forms for each checkbox, with the same name, and a value
of '0'. The checbox, when selected, will send a value of '1' that
overwrites the hidden '0'.

This commit changes just the validation side of the issue; the next
will take care of updating the generated forms