Change the pre-commit verification step
Previously, the code was fixed (php-cs-fixer) automatically, which could
cause merge conflicts with the stashed un-commited changes. Now, the
pre-commit hook simply checks that there are style coding issues and
refuses to commit if there are, saying so. This forces the developer to
stash the changes themselves.

A better approach would be to automatically fix files that do not have
any un-staged changes, but that will maybe be done in the future.