Improve the generate tests
Now the generated files test that the application offers default values
on the create form, and that some tests that would be irrelevant for
some applications are not actually generated. For example, the test to
see if a create form contains default values only makes sense if the
table actually has default values in its schema.