Begin speeding up tests
The Generator tests depend upon actual migrations that must run and be
refreshed every test, which is slow and unnecessary. This commit begins
the process of making the tests not dependent on actual migrations, but
to depend instead on mocked TableInformation (and possibly
DatabaseInformation) instances, which do not require migrations to run
and be cleaned afterwards, with the added benefit that the tests are
less coupled.

This commit changed the ControllerGenerator tests as a start. The
running times for this test suite went from about 1.27 seconds to .48
seconds (a speed-up of 2.6x)!