"Improve" the tests for `TestGenerator`
The tests are a bit improved, they are less coupled by now depending on
an actual database and migration set, but I don't necessarily like how
this code turned out.

I'm now testing the inner methods of the generator, but either the names
of the tests don't match what is actually being tested, or the tests
seem insufficient or inelegant. I will most likely want to revisit this
test case.