Fix XPath of textarea
When the default of a text column is the empty string, the test for the
start values of the create view fails because the XPath

    //textarea[text()='']

Does not return the element '<textarea></textarea>' (because there is no
inner child text-element containing exactly the empty string, since
there is absolutely no child element!). The ".=" operator seems to fix
this, so this

    //textarea[.='']

works.