Remove casting time columns in eloquent models
Laravel does not have a 'time' cast. I was originally storing the time
as a full date and using only the time portion of it, but this
introduces problems when verifying uniqueness. We now completely deal
with this problem by storing a string in the normalized form HH:MM:SS
(padding with 0's where necessary), and retriving the string as is,
without any casting.