Implement a wizard to help with testing
I want the generated code to be valid and working. To test that, I need
to be able to generate an empty laravel project, generate some
migrations, run the autocrud command, and then run php.

The new `wizard` file is able to prepare and setup an empty project that
makes that process more seemless.