Generate gentler migrations to test
Some nouns were tripping up the generator because we were trying to
pluralize an already plural word, etc. I removed (some of) the offending
words from the list of possible nouns.

Additionally, I removed the ability of boolean columns to be unique in
the generated migrations. It doesn't make (much) sense to have them.